// src/app/models/bank-account.model.ts
export class BankAccount {
  accountCountry: string;
  accountName: string;
  accountNumber: string;
  bankName: string;
  branchAddress: string;
  branchName: string;
  clearingCode: string;
  clearingCodeType: string;
  bookedAt: number;
  currency: string;
  id: number;
  swiftCode: string;
  updatedAt: number;
}